requirements.txt
fastapi
uvicorn[standard]
langchain
google-generativeai
pinecone-client
supabase
langchain-google-genai
